+++ addons/test_website/__manifest__.py
@@ -43,9 +43,6 @@ models which only purpose is to run tests.""",
         'web.assets_tests': [
             'test_website/static/tests/tours/*',
         ],
-        'web.qunit_suite_tests': [
-            'test_website/static/tests/*.js',
-        ],
     },
     'license': 'LGPL-3',
 }
