+++ addons/web_unsplash/static/tests/html_field_tests.js
@@ -1,0 +1,7 @@
+/** @odoo-module **/
+
+import { unsplashService } from "@web_unsplash/services/unsplash_service";
+import { mediaDialogServices } from "@web_editor/../tests/html_field_tests";
+
+// update the list of required services to open the media dialog
+mediaDialogServices.unsplash = unsplashService;
